refact: move all common code to base class
