+++ graph_learn/dictionary/base_model.py
@@ -19,14 +19,14 @@ from graph_learn.evaluation import relative_error
 from graph_learn.operators import (
     dictionary_smoothness,
     laplacian_squareform_vec,
-    op_activations_norm,
-    op_weights_norm,
     squared_pdiffs,
 )
 
 
 class GraphDictBase(ABC, BaseEstimator):
     """Graph components learning original method"""
+
+    # INITIALIZATION ###############################################################################
 
     def __init__(
         self,
@@ -186,7 +186,21 @@ class GraphDictBase(ABC, BaseEstimator):
             index=np.arange(self.max_iter),
         )
 
+    # UPDATE ACTIVATIONS ###########################################################################
+
     @abstractmethod
+    def _op_adj_activations(self, weights: NDArray, dualv: NDArray) -> NDArray:
+        """Compute the adjoint of the bilinear inst-degree operator wrt activations
+
+        Args:
+            weights (NDArray): Array of weights of shape (n_components, n_edges)
+            dualv (NDArray): Instantaneous degrees, of shape (n_nodes, n_samples)
+
+        Returns:
+            NDArray: Adjoint activations of shape (n_components, n_samples)
+        """
+        raise NotImplementedError
+
     def _update_activations(
         self,
         sq_pdiffs: NDArray[np.float_],
@@ -196,8 +210,8 @@ class GraphDictBase(ABC, BaseEstimator):
         """Update activations
 
         Args:
-            sq_pdiffs (NDArray[np.float_]): Edgewise squared dstances betweens
-                signal matrix of shape (n_edges, n_nodes)
+            sq_pdiffs (NDArray[np.float_]): Squared pairwise differences of
+                shape (n_samples, n_edges)
             activations (NDArray[np.float_]): Current activations of shape (n_atoms, n_samples)
             dual (NDArray[np.float_]): Dual variable (instantaneous Laplacians)
                 of shape (n_samples, n_nodes, n_nodes)
@@ -205,9 +219,60 @@ class GraphDictBase(ABC, BaseEstimator):
         Returns:
             NDArray[np.float_]: Updated activations of shape (n_atoms, n_samples)
         """
-        raise NotImplementedError
+
+        n_samples = sq_pdiffs.shape[0]
+        # Smoothness
+        step = self.weights_ @ sq_pdiffs.T
+
+        if self.window_size > 1:
+            # average non-overlapping windows
+            step = step.reshape(self.n_atoms, -1, self.window_size).mean(2)
+
+        # L1 regularization
+        step += self.l1_a * self.n_samples_
+
+        if self.log_a > 0:
+            # step -= self.log_a / activations.sum(axis=0, keepdims=True)
+            activations[:, activations.sum(0) < 1e-8] = self.log_a
+
+        if self.l1_diff_a > 0:
+            step -= (
+                self.l1_diff_a
+                * self.n_samples_
+                * (np.diff(np.sign(np.diff(activations, axis=1)), axis=1, prepend=0, append=0))
+            )
+
+        dual_step = self._op_adj_activations(self.weights_, dual)
+
+        # Proximal and gradient step
+        # NOTE: the step might be divided by the operator norm
+        activations = activations - self.step_a / n_samples * (dual_step + step)
+
+        # Projection
+        activations[activations < 0] = 0
+        activations[activations > 1] = 1
+
+        if np.allclose(activations, 0):
+            warn("All activations dropped to 0", UserWarning)
+            activations.fill(self.log_a)
+
+        return activations
+
+    # UPDATE WEIGHTS ###############################################################################
 
     @abstractmethod
+    def _op_adj_weights(self, activations: NDArray, dualv: NDArray) -> NDArray:
+        """Compute the adjoint of the bilinear inst-degree operator wrt weights
+
+        Args:
+            activations (NDArray): Array of activations of shape (n_components, n_samples)
+            dualv (NDArray): Instantaneous degrees, of shape (n_nodes, n_samples)
+
+        Returns:
+            NDArray: Dual weights of shape (n_components, n_edges)
+        """
+        raise NotImplementedError()
+
     def _update_weights(
         self,
         sq_pdiffs: NDArray[np.float_],
@@ -224,7 +289,25 @@ class GraphDictBase(ABC, BaseEstimator):
         Returns:
             NDArray[np.float_]: updated weights of shape (n_atoms, (n_nodes**2 - n_nodes) // 2)
         """
-        raise NotImplementedError
+        n_samples = sq_pdiffs.shape[0]
+
+        # Smoothness
+        step = self.activations_ @ sq_pdiffs
+        step += self.l1_w * self.n_samples_
+
+        if self.ortho_w > 0:
+            step += self.ortho_w * self.n_samples_ * (weights.sum(0, keepdims=True) - weights)
+
+        dual_step = self._op_adj_weights(self.activations_, dual)
+
+        # Proximal update
+        weights = weights - self.step_w / n_samples * (dual_step + step)
+
+        # Projection
+        weights[weights < 0] = 0
+        return weights
+
+    # UPDATE DUAL ##################################################################################
 
     @abstractmethod
     def _update_dual(
@@ -235,19 +318,28 @@ class GraphDictBase(ABC, BaseEstimator):
     ):
         raise NotImplementedError
 
-    def _fit_step(self, sq_pdiffs: NDArray[np.float_]) -> (float, float):
+    def _fit_step(self, sq_pdiffs: NDArray[np.float_]) -> tuple[float, float]:
         # primal update
         # x1 = prox_gx(x - step * (op_adjx(x, dualv) + gradf_x(x, y, gz)), step)
         # y1 = prox_gy(y - step * (op_adjy(y, dualv) + gradf_y(x, y, gz)), step)
-        activations = self._update_activations(sq_pdiffs, self.activations_, dual=self.dual_)
-        weights = self._update_weights(sq_pdiffs, self.weights_, dual=self.dual_)
+
+        activations = np.repeat(
+            self._update_activations(
+                sq_pdiffs, self.activations_[:, :: self.window_size], dual=self.dual_
+            ),
+            repeats=self.window_size,
+            axis=1,
+        )
+        weights = self._update_weights(
+            sq_pdiffs, self.weights_, dual=np.repeat(self.dual_, self.window_size, 1)
+        )
 
         # dual update
         # x_overshoot = 2 * activations - self.activations_
-        # NOTE: I tested overshoot and solutions are either exactly the same, or slightly worse
         self.dual_ = self._update_dual(
-            weights=weights,
-            activations=activations,
+            weights=2 * weights - self.weights_,
+            activations=2 * activations[:, :: self.window_size]
+            - self.activations_[:, :: self.window_size],
             dual=self.dual_,
         )
 
@@ -258,6 +350,8 @@ class GraphDictBase(ABC, BaseEstimator):
         self.weights_ = weights
 
         return a_rel_change, w_rel_change
+
+    # FITTING FUNCTIONS ############################################################################
 
     def _single_fit(
         self,
@@ -310,54 +404,6 @@ class GraphDictBase(ABC, BaseEstimator):
         self.fit_time_ = time() - start
 
         return self
-
-    def score(self, x: NDArray[np.float_], _y=None) -> float:
-        """Compute the negative log-likelihood of the model
-
-        Args:
-            x (NDArray[np.float_]): Design matrix of shape (n_samples, n_nodes)
-            y (None, optional): Ignored. Defaults to None.
-
-        Returns:
-            float: Score of the model
-        """
-        # lw(self.weights_) + 1/self.n_samples * (
-        #     la(self.activations_) + sum_t(
-        #        - np.log(gdet(L_inst)) + x.T @ L_inst @ x
-        #    )
-        # )
-        if np.allclose(self.weights_, 0):
-            warn("All weights are 0, log determinant is undefined", UserWarning)
-            return np.inf
-
-        # sum of L1 norm, orthogonality and smoothness
-        weight_loss = self.l1_w * np.abs(self.weights_).sum() + dictionary_smoothness(  # L1
-            self.activations_, self.weights_, x
-        )
-
-        if self.ortho_w > 0:
-            # gradient
-            weight_loss += self.ortho_w * (
-                np.sum(self.weights_.T @ self.weights_) - np.linalg.norm(self.weights_) ** 2
-            )
-
-        activation_loss = self.l1_a * np.abs(self.activations_).sum()
-        if self.log_a > 0:
-            activation_loss -= self.log_a * np.sum(np.log(self.activations_.sum(0)))
-        if self.l1_diff_a > 0:
-            warn("Check diff loss", UserWarning)
-            activation_loss -= self.l1_diff_a * np.linalg.norm(
-                np.diff(np.sign(np.diff(self.activations_, axis=1)), axis=1, prepend=0, append=0)
-            )
-
-        # Sum of log determinants
-        eigvals = np.linalg.eigvalsh(
-            np.einsum("kc,knm->cnm", self.activations_, laplacian_squareform_vec(self.weights_))
-        )
-        # shape: (n_samples, n_combi) x (n_combi,)
-        loggdet = np.sum(np.log(np.where(eigvals > 0, eigvals, 1)).sum(-1))
-
-        return weight_loss + (activation_loss - loggdet)
 
     def fit(
         self, x: NDArray[np.float_], _y=None, *, callback: Callable[[GraphDictBase, int]] = None
@@ -416,7 +462,87 @@ class GraphDictBase(ABC, BaseEstimator):
         self.activations_ = self.activations_[:, :n_samples]
         return self
 
-    @abstractmethod
+    # PREDICT ######################################################################################
+
     def predict(self, x: NDArray[np.float_]) -> NDArray[np.float_]:
         """Predict activations for a given signal"""
-        raise NotImplementedError
+        n_samples, n_nodes = x.shape
+        if n_nodes != self.n_nodes_:
+            raise ValueError(f"Number of nodes mismatch, got {n_nodes} instead of {self.n_nodes_}")
+
+        activations = self._init_activations(n_samples)
+        dual = self._init_dual(n_samples)
+
+        # op_act_norm = op_activations_norm(laplacian_squareform_vec(self.weights_))
+
+        sq_pdiffs = squared_pdiffs(x)
+
+        for _ in range(self.max_iter):
+            activations_u = np.repeat(
+                self._update_activations(sq_pdiffs, activations[:, :: self.window_size], dual=dual),
+                repeats=self.window_size,
+                axis=1,
+            )
+
+            dual = self._update_dual(
+                self.weights_,
+                2 * activations_u[:, :: self.window_size] - activations[:, :: self.window_size],
+                dual,
+            )
+
+            if np.linalg.norm((activations_u - activations).ravel()) < self.tol:
+                return activations_u
+
+            activations = activations_u
+
+        return activations[:, :n_samples]
+
+    # SCORING ######################################################################################
+
+    def score(self, x: NDArray[np.float_], _y=None) -> float:
+        """Compute the negative log-likelihood of the model
+
+        Args:
+            x (NDArray[np.float_]): Design matrix of shape (n_samples, n_nodes)
+            y (None, optional): Ignored. Defaults to None.
+
+        Returns:
+            float: Score of the model
+        """
+        # lw(self.weights_) + 1/self.n_samples * (
+        #     la(self.activations_) + sum_t(
+        #        - np.log(gdet(L_inst)) + x.T @ L_inst @ x
+        #    )
+        # )
+        if np.allclose(self.weights_, 0):
+            warn("All weights are 0, log determinant is undefined", UserWarning)
+            return np.inf
+
+        # sum of L1 norm, orthogonality and smoothness
+        weight_loss = self.l1_w * np.abs(self.weights_).sum() + dictionary_smoothness(  # L1
+            self.activations_, self.weights_, x
+        )
+
+        if self.ortho_w > 0:
+            # gradient
+            weight_loss += self.ortho_w * (
+                np.sum(self.weights_.T @ self.weights_) - np.linalg.norm(self.weights_) ** 2
+            )
+
+        activation_loss = self.l1_a * np.abs(self.activations_).sum()
+        if self.log_a > 0:
+            activation_loss -= self.log_a * np.sum(np.log(self.activations_.sum(0)))
+        if self.l1_diff_a > 0:
+            warn("Check diff loss", UserWarning)
+            activation_loss -= self.l1_diff_a * np.linalg.norm(
+                np.diff(np.sign(np.diff(self.activations_, axis=1)), axis=1, prepend=0, append=0)
+            )
+
+        # Sum of log determinants
+        eigvals = np.linalg.eigvalsh(
+            np.einsum("kc,knm->cnm", self.activations_, laplacian_squareform_vec(self.weights_))
+        )
+        # shape: (n_samples, n_combi) x (n_combi,)
+        loggdet = np.sum(np.log(np.where(eigvals > 0, eigvals, 1)).sum(-1))
+
+        return weight_loss + (activation_loss - loggdet)
